refactor: extract turn orchestration
